+++ Gemfile
@@ -4,9 +4,6 @@ source 'http://rubygems.org'
 
 gemspec
 
-gem 'aruba', '~> 1.0.2'
-gem 'byebug', '~> 11.1.0'
-gem 'pry', '~> 0.13.0'
 gem 'rake', '~> 12.0'
 gem 'rspec', '~> 3.0'
 gem 'rubocop', '~> 0.88'
