+++ spec/shared_examples/shared_command_with_times_spec.rb
