+++ spec/spec_helper.rb
@@ -2,40 +2,16 @@
 
 require 'bundler/setup'
 require 'ruby_jard'
-require_relative './helpers/shared_command_with_times_spec'
 
-##
-# Config to test pry-related commands. Copy from pry and pry-byebug test base
-Pry.config.color = false
-Pry.config.pager = false
-Pry.config.correct_indent = false
-
-def redirect_pry_io(new_in, new_out = StringIO.new)
-  old_in = Pry.input
-  old_out = Pry.output
-  Pry.input = new_in
-  Pry.output = new_out
-  begin
-    yield
-  ensure
-    Pry.input = old_in
-    Pry.output = old_out
+def require_relative_folder(pattern)
+  pattern = File.expand_path(File.join(File.dirname(__FILE__), pattern))
+  Dir[pattern].sort.each do |file|
+    require file
   end
 end
 
-class InputTester
-  def initialize(*actions)
-    @actions = actions
-  end
-
-  def add(*actions)
-    @actions += actions
-  end
-
-  def readline(*)
-    @actions.shift
-  end
-end
+require_relative_folder('./helpers/**/*')
+require_relative_folder('./shared_examples/**/*')
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
